get rid of warnings, some fixes for nonterm

+++ src/analysis/prune.cpp
@@ -47,7 +47,7 @@ bool Pruning::compareRules(const Rule &a, const Rule &b, bool compareRhss) {
 
     // All right-hand sides have to match exactly
     if (compareRhss) {
-        for (int i=0; i < a.rhsCount(); ++i) {
+        for (unsigned int i=0; i < a.rhsCount(); ++i) {
             const UpdateMap &updateA = a.getUpdate(i);
             const UpdateMap &updateB = b.getUpdate(i);
 
@@ -64,7 +64,7 @@ bool Pruning::compareRules(const Rule &a, const Rule &b, bool compareRhss) {
     }
 
     // Guard has to be fully equal (including the ordering)
-    for (int i=0; i < guardA.size(); ++i) {
+    for (unsigned int i=0; i < guardA.size(); ++i) {
         if (!guardA[i].is_equal(guardB[i])) return false;
     }
     return true;
@@ -146,9 +146,9 @@ bool Pruning::pruneParallelRules(ITSProblem &its) {
             if (parallel.size() > Config::Prune::MaxParallelRules) {
                 PriorityQueue queue(comp);
 
-                for (int i=0; i < parallel.size(); ++i) {
+                for (unsigned int i=0; i < parallel.size(); ++i) {
                     // alternating iteration (idx=0,n-1,1,n-2,...) that might avoid choosing similar edges
-                    int idx = (i % 2 == 0) ? i/2 : parallel.size()-1-i/2;
+                    unsigned long idx = (i % 2 == 0) ? i/2 : parallel.size()-1-i/2;
 
                     TransIdx ruleIdx = parallel[idx];
                     const Rule &rule = its.getRule(parallel[idx]);
@@ -166,7 +166,7 @@ bool Pruning::pruneParallelRules(ITSProblem &its) {
 
                 // Keep only the top elements of the queue
                 set<TransIdx> keep;
-                for (int i=0; i < Config::Prune::MaxParallelRules; ++i) {
+                for (unsigned int i=0; i < Config::Prune::MaxParallelRules; ++i) {
                     keep.insert(get<0>(queue.top()));
                     queue.pop();
                 }
